% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipify.R
\name{get_ip}
\alias{get_ip}
\title{Get IP}
\usage{
get_ip(max_tries = ipify.env$MAX_TRIES, ...)
}
\arguments{
\item{max_tries}{The number of times to retry the ipify service on failure. Defaults to 3.}

\item{...}{parameter = Format. What you  want the service to return. Valid options = c("text", "json", "jsonp")}
}
\value{
The public IP address of this machine as a plain text string,
}
\description{
Query the ipify service (https://www.ipify.org) to retrieve this machine's
public IP address.
}
