% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipify.R
\name{get_ip_response}
\alias{get_ip_response}
\title{Get a response object from ipify}
\usage{
get_ip_response(url = ipify.env$API_URL, ...)
}
\arguments{
\item{url}{The base url of the ipify service.}

\item{...}{parameter = Format. What you  want the service to return. Valid options = c("text", "json", "jsonp")}
}
\description{
This is a private, non-exported function
}
