% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{ipify.env}
\alias{ipify.env}
\title{Set up the environment used to store setting information}
\format{An object of class \code{environment} of length 3.}
\usage{
ipify.env
}
\description{
Set up the environment used to store setting information
}
\keyword{datasets}
